requirements.txt
annotated-types==0.7.0
anthropic==0.75.0

anyio==4.8.0
certifi==2025.11.12
distro==1.9.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-auth==2.45.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.9.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.2.6
openai==2.2.0
pandas==2.3.3
path==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sniffio==1.3.1
streamlit==1.52.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1